plot.cv.hqreg {hqreg}	R Documentation

Plot the cross-validation curve for a "cv.hqreg" object

Description

Plot the cross-validation curve for a "cv.hqreg" object against the lambda values used, along with standard error bars.

Usage

## S3 method for class 'cv.hqreg'
plot(x, log.l = TRUE, nvars = TRUE, ...)

Arguments

x
A "cv.hqreg" object.

log.l
Should log(lambda) be used instead of lambda for X-axis? Default is TRUE.

nvars
If TRUE (the default), places an axis on top of the plot denoting the number of variables with nonzero coefficients at each lambda.

...
Other graphical parameters to plot

Details

Produces a plot of mean cv errors at each lambda along with upper and lower standard error bars.

Author(s)

Congrui Yi <congrui-yi@uiowa.edu>

References

Yi, C. and Huang, J. (2016) Semismooth Newton Coordinate Descent Algorithm for Elastic-Net Penalized Huber Loss Regression and Quantile Regression, https://arxiv.org/abs/1509.02957
Journal of Computational and Graphical Statistics, accepted in Nov 2016
http://www.tandfonline.com/doi/full/10.1080/10618600.2016.1256816

See Also

hqreg, cv.hqreg

Examples

X = matrix(rnorm(1000*100), 1000, 100)
beta = rnorm(10)
eps = 4*rnorm(1000)
y = drop(X[,1:10] %*% beta + eps)
cv = cv.hqreg(X, y, seed = 123)
plot(cv)

[Package hqreg version 1.4 Index]